+++ api/v1/src/models/Questionnaire.generated.ts
@@ -227,7 +227,9 @@ export type LogicOperator =
   | LogicOr
   | LogicIn
   | LogicGreater
-  | LogicLess;
+  | LogicLess
+  | LogicConvertToDateString
+  | Round;
 export type LogicConstant = number | string | boolean;
 /**
  * Type of the question.
@@ -388,6 +390,12 @@ export interface LogicGreater {
 }
 export interface LogicLess {
   "<": [LogicExpression, LogicExpression];
+}
+export interface LogicConvertToDateString {
+  convert_to_date_string: [LogicExpression, LogicExpression];
+}
+export interface Round {
+  round: LogicExpression | [LogicExpression];
 }
 export interface LogicVariable {
   var: string;