+++ covquestions-editor-app/src/App.tsx
@@ -37,6 +37,7 @@ import { useSelector } from "react-redux";
 import { getAllQuestionnaires, getQuestionnaireByIdVersionAndLanguage } from "./api/api-client";
 import { QuestionnaireBaseData } from "./models/QuestionnairesList";
 import { SettingSelection } from "./components/questionnaireSelection/SettingSelection";
+import { getQueryParams, setQueryParams } from "./utils/queryParams";
 import ReactMarkdown from "react-markdown";
 import { generatedInstructionsFromMarkdown } from "./generatedInstructionsFromMarkdown";
 
@@ -154,6 +155,8 @@ export const App: React.FC = () => {
       currentQuestionnaireSelection.version !== undefined &&
       currentQuestionnaireSelection.language !== undefined
     ) {
+      setQueryParams(currentQuestionnaireSelection);
+
       getQuestionnaireByIdVersionAndLanguage(
         currentQuestionnaireSelection.id,
         currentQuestionnaireSelection.version,
@@ -185,6 +188,12 @@ export const App: React.FC = () => {
       setExecutedQuestionnaire(currentQuestionnaire.questionnaire);
     }
   }, [currentQuestionnaire, hasAnyError]);
+
+  // Select Questionnaire that is saved in the query params
+  const querySelection: QuestionnaireSelection = getQueryParams();
+  if (querySelection.id != null && currentQuestionnaireSelection.id == null) {
+    setCurrentQuestionnaireSelection(querySelection);
+  }
 
   return (
     <ThemeProvider theme={theme}>