+++ api/v1/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.2](https://github.com/CovOpen/CovQuestions/compare/cov-question-api-v1@1.3.1...cov-question-api-v1@1.3.2) (2021-04-11)
+
+**Note:** Version bump only for package cov-question-api-v1
+
+
+
+
+
+## [1.3.1](https://github.com/CovOpen/CovQuestions/compare/cov-question-api-v1@1.3.0...cov-question-api-v1@1.3.1) (2021-04-09)
+
+
+### Bug Fixes
+
+* trigger file update in api ([f471968](https://github.com/CovOpen/CovQuestions/commit/f4719687954c6fa9278e3878141cd5a9e1eae872))
+
+
+
+
+
+# [1.3.0](https://github.com/CovOpen/CovQuestions/compare/cov-question-api-v1@1.2.0...cov-question-api-v1@1.3.0) (2021-04-08)
+
+
+### Features
+
+* add rounding functions ([1233a71](https://github.com/CovOpen/CovQuestions/commit/1233a7196d9232ee72e86978e3a8b82c201ab820))
+
+
+
+
+
+# [1.2.0](https://github.com/CovOpen/CovQuestions/compare/cov-question-api-v1@1.1.0...cov-question-api-v1@1.2.0) (2021-04-05)
+
+
+### Features
+
+* add date conversion to schema ([cc71467](https://github.com/CovOpen/CovQuestions/commit/cc7146717f2785390a48df63ff69328fea65b2f6))
+
+
+
+
+
 # [1.1.0](https://github.com/CovOpen/CovQuestions/compare/cov-question-api-v1@1.0.7...cov-question-api-v1@1.1.0) (2021-03-30)